Moved all libraries (netty, LibGDX, etc) into the libs/ directory of the source code.
Began porting to LibGDX - breaking pretty much everything. The port now
has window that will open with the standard grass background, and the
player does spawn and accept input, however the player's sprite isn't
visible. 

Most sprite functions (etc) have been replaced with placeholders so that
the project will compile, as sprites are now handled by a member class
in the Player class, which in turn extends the Sprite class of LibGDX. 

Some Netty methods stopped working correctly with the minor increment to
netty version - this needs addressing. 

Lots of files that haven't been ported, or have been ported and old
versions still exist are present in this commit!